+++ src/geocode/IDrawableLocation.ts
@@ -2,7 +2,7 @@ import { LatLng } from "leaflet";
 /**
  * Interface representing a drawable Location (e.g. Marker)
  */
-export default interface IDrawablePoint{
+export interface IDrawablePoint{
     name:string,
     position: LatLng,
 }
