+++ src/geocode/IDrawablePolyLine.ts
@@ -2,7 +2,7 @@ import { LatLng } from "leaflet";
 /**
  * Interface representing a polyline wich can be drawe on a map
  */
-export default interface IDrawablePolyLine{
+export interface IDrawablePolyLine{
     /**
      * Orderd list of waypoints
      */
