+++ src/geocode/IDrawableStreet.ts
@@ -1,10 +1,10 @@
 import { LatLng, PathOptions } from "leaflet";
-import IDrawablePolyLine from "./IDrawablePolyLine";
+import { IDrawablePolyLine } from "./IDrawablePolyLine";
 
 /**
  * Interface containing all values needed for the street element to display it
  */
-export default interface IDrawableStreet{
+export interface IDrawableStreet{
     name:string,
     pathOption?: PathOptions,
     center:LatLng,
