+++ src/learningFileHandling/index.ts
@@ -1,0 +1,4 @@
+export * from "./FileHandler"
+export * from "./FileSelector"
+export * from "./ILearningFile"
+export * from "./LearningFile"
